Add files via upload
# ICD 11 Data (French)

This RData file contains the data obtained by scraping the International Classification of Diseases (ICD) 11th Edition in the French language. The data was extracted using the corresponding R script available in the repository.

## Contents
The file includes structured data related to medical codes and classifications in the French language. The data is organized based on the ICD 11 hierarchy, including chapters, sections, and detailed entities.

## Objects
### 1. ICD 11 Entities List
This is a list containing structured data related to medical codes and classifications in the French language. The data is organized based on the ICD 11 hierarchy, including chapters, sections, and detailed entities.

### 2. Flattened Data Frame
A data frame obtained by flattening the list of ICD 11 entities. This provides a tabular representation of the hierarchical data for easier analysis and integration into various R-based projects.

## Usage
This dataset can be utilized for research, analysis, or any other purposes related to medical classifications and coding. It is provided in RData format for easy integration into R-based projects.
- The ICD 11 Entities List can be explored for detailed hierarchical information.
- The Flattened Data Frame is suitable for research, analysis, or any other purposes related to medical classifications and coding.
## Note
Please adhere to the terms of use and licensing agreements of the ICD 11 API when using this data.

Feel free to contribute or provide feedback!

[GitHub Repository](https://github.com/EpidemiologieSantePublique/icd11-scarper)